Add weights to the flex-items for their widths
We added three blank footer flex containers at the bottom and a related piece of CSS. The idea is to use flex property on one of these footer items. By providing value of 2 to flex property of the class footer-one, we are able to make the width of the first flex container twice as the size of the other containers. The container kind of grows twice as fast and occupies the available space. If we apply this value 2 to flex container one and two, they will also grow twice the width. The only thing is that the width will decrease compared to when the first flex container alone is set to 2. Because now the available space is less.