Add auto-margin to the menu
In this interesting property usage, we removed the additional div tag with links class. We had created it with div tag to keep the signup and login links at one side. We removed that div tag and got the same behaviour by writing a CSS property for signup class with auto property. This gave us the same behaviour as with the previous div tag.